Move mockup pages into pages/mockups/
