+++ packages/playground/src/pages/mockups/ContactDetail/ContactDetail.module.scss
