+++ packages/playground/src/pages/mockups/Dashboard/Dashboard.tsx
@@ -5,7 +5,7 @@ import { Cluster } from '@eocrm/design-system';
 import { Avatar } from '@eocrm/design-system';
 import { Badge } from '@eocrm/design-system';
 import { Button } from '@eocrm/design-system';
-import { contacts, deals } from '../../data/mock';
+import { contacts, deals } from '../../../data/mock';
 import styles from './Dashboard.module.scss';
 
 const stats = [
