+++ packages/playground/src/pages/mockups/Deals/Deals.tsx
@@ -4,7 +4,7 @@ import { Badge } from '@eocrm/design-system';
 import { Button } from '@eocrm/design-system';
 import { Cluster } from '@eocrm/design-system';
 import { Stack } from '@eocrm/design-system';
-import { dealStages, deals, type DealStage } from '../../data/mock';
+import { dealStages, deals, type DealStage } from '../../../data/mock';
 import styles from './Deals.module.scss';
 
 export function Deals() {
